Move some code into airquality package
