--- backend/src/main/scala/com/github/sophiecollard/domain/clients/LondonAirClientAlgebra.scala
@@ -1,10 +1,0 @@
-package com.github.sophiecollard.domain.clients
-
-import com.github.sophiecollard.domain.error.AppError
-import com.github.sophiecollard.domain.model.DailyAirQualityIndex
-
-trait LondonAirClientAlgebra[F[_]] {
-
-  def getDailyAirQualityIndex: F[Either[AppError, DailyAirQualityIndex]]
-
-}
